# packages/studio-ui/public/openapi/acton-localnet-control.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Acton Localnet Control API",
    "version": "1.0.0",
    "description": "Local development control endpoints exposed by the Acton localnet server. These routes are intended for local tooling. When localnet is started with --require-auth, pass the printed token as Authorization: Bearer <token> or X-API-Key: <token>."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5411",
      "description": "Default localnet server"
    }
  ],
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "APIKeyHeader": []
    }
  ],
  "paths": {
    "/acton_fundAccount": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "fundAccount",
        "summary": "Fund an account",
        "description": "Queues a local faucet transfer to credit an address. The transfer is included on the next automatic or manually mined block.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundAccountRequest"
              },
              "example": {
                "address": "0:2222222222222222222222222222222222222222222222222222222222222222",
                "amount": 1000000000
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/FundAccountResponse"
          }
        }
      }
    },
    "/acton_fundJetton": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "fundJetton",
        "summary": "Mint jettons to an account",
        "description": "Finds the jetton master in localnet state, verifies that it is mintable and has an admin, converts the display amount using the decimals from jetton metadata, preflights the legacy 0x15 and current 0x642b7d07 mint layouts, and queues the uniquely accepted message for the next block.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundJettonRequest"
              },
              "example": {
                "address": "0:2222222222222222222222222222222222222222222222222222222222222222",
                "jetton_master": "0:3333333333333333333333333333333333333333333333333333333333333333",
                "amount": "12.5"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SendInternalMessageResponse"
          }
        }
      }
    },
...
    "/acton_getVerifiedSource": {
      "get": {
        "tags": ["Localnet Control"],
        "operationId": "getVerifiedSource",
        "summary": "Get verified contract source",
        "description": "Proxies Acton verifier source lookup by address or code hash and returns the verified source bundle when the contract is confirmed.",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Address"
            }
          },
          {
            "name": "code_hash",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Hash"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/VerifiedSourceResponse"
          }
        }
      }
    },
...
    "/acton_loadState": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "loadState",
        "summary": "Load local state",
        "description": "Replaces the current node state from the request body and clears all checkpoints.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateFile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/NullResponse"
          }
        }
      }
    },
    "/acton_createCheckpoint": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "createCheckpoint",
        "summary": "Create checkpoint",
        "description": "Creates a named in-memory checkpoint for the current localnet state.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckpointRequest"
              },
              "example": {
                "name": "before-upgrade",
                "force": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CheckpointResponse"
          }
        }
      }
    },
...
    "/acton_restoreCheckpoint": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "restoreCheckpoint",
        "summary": "Restore checkpoint",
        "description": "Restores a checkpoint without deleting it or checkpoints created after it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointRequest"
              },
              "example": {
                "name": "before-upgrade"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CheckpointResponse"
          }
        }
      }
    },
    "/acton_deleteCheckpoint": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "deleteCheckpoint",
        "summary": "Delete checkpoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckpointRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CheckpointResponse"
          }
        }
      }
    },
...
    "/acton_exportCheckpoint": {
      "get": {
        "tags": ["Localnet Control"],
        "operationId": "exportCheckpoint",
        "summary": "Download checkpoint",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Checkpoint state file.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateFile"
                }
              }
            }
          }
        }
      }
    },
    "/acton_importCheckpoint": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "importCheckpoint",
        "summary": "Import checkpoint",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StateFile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/CheckpointResponse"
          }
        }
      }
    },
    "/acton_setConfig": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "setConfig",
        "summary": "Replace blockchain config",
        "description": "Validates and replaces the full blockchain config dictionary from a base64-encoded BOC, then commits the change in a config-only block.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetConfigRequest"
              },
              "example": {
                "config": "te6ccgEC..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SetConfigResponse"
          }
        }
      }
    },
    "/acton_setShardAccount": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "setShardAccount",
        "summary": "Set a ShardAccount",
        "description": "Replaces the selected account state with a base64-encoded ShardAccount BOC.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetShardAccountRequest"
              },
              "example": {
                "address": "0:2222222222222222222222222222222222222222222222222222222222222222",
                "shard_account": "te6ccgEBAQEA..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/NullResponse"
          }
        }
      }
    },
    "/acton_changeAccountState": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "changeAccountState",
        "summary": "Change account state",
        "description": "Changes the selected local account state to nonexist, uninit, or frozen. When state is frozen with source=current, set mine=false to queue the synthetic freeze transaction for the next mined block instead of committing a block immediately.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAccountStateRequest"
              },
              "example": {
                "address": "0:2222222222222222222222222222222222222222222222222222222222222222",
                "mine": false,
                "state": {
                  "type": "frozen",
                  "source": "current"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/NullResponse"
          }
        }
      }
    },
    "/acton_sendInternalMessage": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "sendInternalMessage",
        "summary": "Send a raw internal message",
        "description": "Injects a base64-encoded internal message BOC into the local internal queue. TonCenter-compatible message endpoints accept external-in messages only.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBocRequest"
              },
              "example": {
                "boc": "te6ccgEBAQEA..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SendInternalMessageResponse"
          }
        }
      }
    },
...
    "/acton_mine": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "mineBlocks",
        "summary": "Mine localnet blocks",
        "description": "Manually mines queued and/or empty localnet blocks. When the request body is omitted or blocks is not provided, one block is mined.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MineBlocksRequest"
              },
              "example": {
                "blocks": 3
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MineBlocksResponse"
          }
        }
      }
    },
    "/acton_increaseTime": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "increaseTime",
        "summary": "Increase virtual time",
        "description": "Adds seconds to the localnet virtual clock. Future blocks use this adjusted time until it is changed again.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncreaseTimeRequest"
              },
              "example": {
                "seconds": 3600
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TimeControlResponse"
          }
        }
      }
    },
    "/acton_setTime": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "setTime",
        "summary": "Set virtual time",
        "description": "Sets the localnet virtual clock to a Unix timestamp. The timestamp must not be lower than the latest mined block time.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTimeRequest"
              },
              "example": {
                "timestamp": 1710000000
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TimeControlResponse"
          }
        }
      }
    },
    "/acton_setNextBlockTimestamp": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "setNextBlockTimestamp",
        "summary": "Set next block timestamp",
        "description": "Sets a one-shot Unix timestamp for the next mined block. The timestamp is consumed by the next automatic or manually mined block and must not be lower than the latest mined block time.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetNextBlockTimestampRequest"
              },
              "example": {
                "timestamp": 1710000600
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/TimeControlResponse"
          }
        }
      }
    },
    "/acton_setNetworkConditions": {
      "post": {
        "tags": ["Localnet Control"],
        "operationId": "setNetworkConditions",
        "summary": "Set simulated network conditions",
        "description": "Updates simulated network conditions while the localnet server is running and returns the updated values. Set response_delay_ms to 0 to disable response delay. Current values are also available from /acton_nodeInfo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetNetworkConditionsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/NetworkConditionsResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Localnet API token printed by `acton localnet start --require-auth`."
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative localnet API token header for TonCenter-compatible clients."
      }
    },
...
    "responses": {
...
      "FundAccountResponse": {
        "description": "Successful faucet response.",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/components/schemas/FundAccountResultWrapper"
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
    },
    "schemas": {
      "Address": {
        "type": "string",
        "description": "TON address in raw or user-friendly format."
      },
...
      "FundAccountRequest": {
        "type": "object",
        "required": ["address", "amount"],
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "amount": {
            "type": "integer",
            "minimum": 0,
            "description": "Amount in nanograms."
          }
        }
      },
      "FundJettonRequest": {
        "type": "object",
        "required": ["address", "jetton_master", "amount"],
        "properties": {
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "jetton_master": {
            "$ref": "#/components/schemas/Address"
          },
          "amount": {
            "type": "string",
            "pattern": "^[0-9]+(?:\\.[0-9]*)?$",
            "description": "Positive display-unit amount. The server reads decimals from the jetton master metadata and defaults to 9 when the metadata value is missing or invalid."
          }
        }
      },
...
    }
  }
}